+++ Tests/EntityFrameworkCore.Repositories.TestApplication/EntityFrameworkCore.Repositories.TestApplication.Infrastructure/Repositories/AggregateRoot1Repository.cs
@@ -287,7 +287,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params0_returnsD_collection0_schemaName0")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -298,28 +298,28 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params0_returnsD_collection1_schemaName0(CancellationToken cancellationToken = default)
+        public async Task<List<SpResult>> Sp_params0_returnsD_collection1_schemaName0(CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params0_returnsD_collection1_schemaName0")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params0_returnsD_collection1_schemaName1(CancellationToken cancellationToken = default)
+        public async Task<List<SpResult>> Sp_params0_returnsD_collection1_schemaName1(CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -329,7 +329,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await GetSet()
                 .FromSqlInterpolated($"EXECUTE sp_params0_returnsE_collection0_schemaName0")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -340,28 +340,28 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await GetSet()
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params0_returnsE_collection1_schemaName0(CancellationToken cancellationToken = default)
+        public async Task<List<AggregateRoot1>> Sp_params0_returnsE_collection1_schemaName0(CancellationToken cancellationToken = default)
         {
             var results = await GetSet()
                 .FromSqlInterpolated($"EXECUTE sp_params0_returnsE_collection1_schemaName0")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params0_returnsE_collection1_schemaName1(CancellationToken cancellationToken = default)
+        public async Task<List<AggregateRoot1>> Sp_params0_returnsE_collection1_schemaName1(CancellationToken cancellationToken = default)
         {
             var results = await GetSet()
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -383,7 +383,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params1_returnsD_collection0_schemaName0 {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -396,32 +396,32 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params1_returnsD_collection1_schemaName0(
+        public async Task<List<SpResult>> Sp_params1_returnsD_collection1_schemaName0(
             string param1,
             CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params1_returnsD_collection1_schemaName0 {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params1_returnsD_collection1_schemaName1(
+        public async Task<List<SpResult>> Sp_params1_returnsD_collection1_schemaName1(
             string param1,
             CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -433,7 +433,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await GetSet()
                 .FromSqlInterpolated($"EXECUTE sp_params1_returnsE_collection0_schemaName0 {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -446,32 +446,32 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await GetSet()
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params1_returnsE_collection1_schemaName0(
+        public async Task<List<AggregateRoot1>> Sp_params1_returnsE_collection1_schemaName0(
             string param1,
             CancellationToken cancellationToken = default)
         {
             var results = await GetSet()
                 .FromSqlInterpolated($"EXECUTE sp_params1_returnsE_collection1_schemaName0 {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params1_returnsE_collection1_schemaName1(
+        public async Task<List<AggregateRoot1>> Sp_params1_returnsE_collection1_schemaName1(
             string param1,
             CancellationToken cancellationToken = default)
         {
             var results = await GetSet()
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -498,7 +498,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params2_returnsD_collection0_schemaName0 {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -512,13 +512,13 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params2_returnsD_collection1_schemaName0(
+        public async Task<List<SpResult>> Sp_params2_returnsD_collection1_schemaName0(
             string param1,
             string param2,
             CancellationToken cancellationToken = default)
@@ -526,12 +526,12 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params2_returnsD_collection1_schemaName0 {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params2_returnsD_collection1_schemaName1(
+        public async Task<List<SpResult>> Sp_params2_returnsD_collection1_schemaName1(
             string param1,
             string param2,
             CancellationToken cancellationToken = default)
@@ -539,7 +539,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -552,7 +552,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await GetSet()
                 .FromSqlInterpolated($"EXECUTE sp_params2_returnsE_collection0_schemaName0 {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -566,13 +566,13 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await GetSet()
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params2_returnsE_collection1_schemaName0(
+        public async Task<List<AggregateRoot1>> Sp_params2_returnsE_collection1_schemaName0(
             string param1,
             string param2,
             CancellationToken cancellationToken = default)
@@ -580,12 +580,12 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await GetSet()
                 .FromSqlInterpolated($"EXECUTE sp_params2_returnsE_collection1_schemaName0 {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params2_returnsE_collection1_schemaName1(
+        public async Task<List<AggregateRoot1>> Sp_params2_returnsE_collection1_schemaName1(
             string param1,
             string param2,
             CancellationToken cancellationToken = default)
@@ -593,7 +593,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await GetSet()
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }