+++ Tests/EntityFrameworkCore.Repositories.TestApplication/EntityFrameworkCore.Repositories.TestApplication.Infrastructure/Repositories/CustomRepository.cs
@@ -29,7 +29,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params0_returnsD_collection0_schemaName0")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -40,28 +40,28 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params0_returnsD_collection1_schemaName0(CancellationToken cancellationToken = default)
+        public async Task<List<SpResult>> Sp_params0_returnsD_collection1_schemaName0(CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params0_returnsD_collection1_schemaName0")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params0_returnsD_collection1_schemaName1(CancellationToken cancellationToken = default)
+        public async Task<List<SpResult>> Sp_params0_returnsD_collection1_schemaName1(CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -71,7 +71,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE sp_params0_returnsE_collection0_schemaName0")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -82,28 +82,28 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params0_returnsE_collection1_schemaName0(CancellationToken cancellationToken = default)
+        public async Task<List<AggregateRoot1>> Sp_params0_returnsE_collection1_schemaName0(CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE sp_params0_returnsE_collection1_schemaName0")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params0_returnsE_collection1_schemaName1(CancellationToken cancellationToken = default)
+        public async Task<List<AggregateRoot1>> Sp_params0_returnsE_collection1_schemaName1(CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -125,7 +125,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params1_returnsD_collection0_schemaName0 {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -138,32 +138,32 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params1_returnsD_collection1_schemaName0(
+        public async Task<List<SpResult>> Sp_params1_returnsD_collection1_schemaName0(
             string param1,
             CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params1_returnsD_collection1_schemaName0 {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params1_returnsD_collection1_schemaName1(
+        public async Task<List<SpResult>> Sp_params1_returnsD_collection1_schemaName1(
             string param1,
             CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -175,7 +175,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE sp_params1_returnsE_collection0_schemaName0 {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -188,32 +188,32 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params1_returnsE_collection1_schemaName0(
+        public async Task<List<AggregateRoot1>> Sp_params1_returnsE_collection1_schemaName0(
             string param1,
             CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE sp_params1_returnsE_collection1_schemaName0 {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params1_returnsE_collection1_schemaName1(
+        public async Task<List<AggregateRoot1>> Sp_params1_returnsE_collection1_schemaName1(
             string param1,
             CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -240,7 +240,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params2_returnsD_collection0_schemaName0 {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -254,13 +254,13 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params2_returnsD_collection1_schemaName0(
+        public async Task<List<SpResult>> Sp_params2_returnsD_collection1_schemaName0(
             string param1,
             string param2,
             CancellationToken cancellationToken = default)
@@ -268,12 +268,12 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE sp_params2_returnsD_collection1_schemaName0 {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<SpResult>> Sp_params2_returnsD_collection1_schemaName1(
+        public async Task<List<SpResult>> Sp_params2_returnsD_collection1_schemaName1(
             string param1,
             string param2,
             CancellationToken cancellationToken = default)
@@ -281,7 +281,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await _dbContext.SpResults
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -294,7 +294,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE sp_params2_returnsE_collection0_schemaName0 {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -308,13 +308,13 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params2_returnsE_collection1_schemaName0(
+        public async Task<List<AggregateRoot1>> Sp_params2_returnsE_collection1_schemaName0(
             string param1,
             string param2,
             CancellationToken cancellationToken = default)
@@ -322,12 +322,12 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE sp_params2_returnsE_collection1_schemaName0 {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
 
-        public async Task<IReadOnlyCollection<AggregateRoot1>> Sp_params2_returnsE_collection1_schemaName1(
+        public async Task<List<AggregateRoot1>> Sp_params2_returnsE_collection1_schemaName1(
             string param1,
             string param2,
             CancellationToken cancellationToken = default)
@@ -335,7 +335,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await _dbContext.AggregateRoot1s
                 .FromSqlInterpolated($"EXECUTE dbo.sp_custom {param1}, {param2}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }