+++ Tests/EntityFrameworkCore.Repositories.TestApplication/EntityFrameworkCore.Repositories.TestApplication.Infrastructure/Repositories/MappableStoredProcs/MappableStoredProcRepository.cs
@@ -31,7 +31,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.EntityRecords
                 .FromSqlInterpolated($"EXECUTE GetEntityById {id}")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return result;
@@ -75,12 +75,12 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             await _dbContext.Database.ExecuteSqlInterpolatedAsync($"EXECUTE CreateEntity {id}, {name}", cancellationToken);
         }
 
-        public async Task<IReadOnlyCollection<EntityRecord>> GetEntities(CancellationToken cancellationToken = default)
+        public async Task<List<EntityRecord>> GetEntities(CancellationToken cancellationToken = default)
         {
             var results = await _dbContext.EntityRecords
                 .FromSqlInterpolated($"EXECUTE GetEntities")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return results;
         }
@@ -109,7 +109,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = (await _dbContext.MappedSpResultItems
                 .FromSqlInterpolated($"EXECUTE MappedSp {paramSomething1WithAltName}, {paramName2WithAltName}, {outputParam1Parameter} OUTPUT, {outputParam2Parameter} OUTPUT")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken))
+                .ToListAsync(cancellationToken))
                 .Single();
 
             return new MappedSpResult(result: result, simpleString: (string)outputParam2Parameter.Value);
@@ -139,7 +139,7 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var results = await _dbContext.MappedSpResultItems
                 .FromSqlInterpolated($"EXECUTE MappedSpCollection {paramRandom2WithAltName}, {paramElse1WithAltName}, {outputParam2Parameter} OUTPUT, {outputParam1Parameter} OUTPUT")
                 .IgnoreQueryFilters()
-                .ToArrayAsync(cancellationToken);
+                .ToListAsync(cancellationToken);
 
             return new MappedSpResultCollection(result: results, simpleString: (string)outputParam2Parameter.Value);
         }
@@ -159,6 +159,16 @@ namespace EntityFrameworkCore.Repositories.TestApplication.Infrastructure.Reposi
             var result = await _dbContext.ExecuteScalarAsync<int>("EXECUTE MappedSpReturningScalar @param1", param1Parameter);
 
             return result;
+        }
+
+        public async Task<List<MappedSpCollectionPassthroughResult>> MappedOperationCollectionPassthrough(CancellationToken cancellationToken = default)
+        {
+            var results = await _dbContext.MappedSpCollectionPassthroughResults
+                .FromSqlInterpolated($"EXECUTE MappedSpCollectionPassthrough")
+                .IgnoreQueryFilters()
+                .ToListAsync(cancellationToken);
+
+            return results;
         }
     }
 }