+++ Modules/Intent.Modules.EntityFrameworkCore.Repositories/release-notes.md
@@ -2,6 +2,7 @@
 
 - Fixed: Stored procedure implementations would generate incorrect code if parameter input names differed from those of the containing method.
 - Fixed: Repository method implementations for mapped stored procedures returning scalar values would have an incorrect return statement generated.
+- Fixed: Updated stored procedure implementations to use `.ToListAsync(...)` and to always return `List<T>` to align with other repository operations as well as the calling operations which typically return `List<T>` as well.
 
 ### Version 4.7.13